pyproject.toml
[tool.poetry]
name = "dz"
version = "0.1.0"
description = ""
authors = ["ViktorAvdeev <flystrike1994@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.0"
mypy = "^1.10.1"
black = "^24.4.2"
isort = "^5.13.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = ['venv', 'env', '.venv']


[tool.isort]
line_length = 119


[tool.black]
line-length = 119
exclude = '''
(
  /(       # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.mypy_cache
    | \.venv
    | venv
    | env
  )/
)
'''


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
